Enable exceptions by on input/output streams.
Motivating example:

Until now, the example `lz4_compress` application would silently
ignore errors like `no space left on device` (ENOSP) when compressing
continuing to churn through the input data and writing to the
output file even though every write fails (as evidenced by `strace`)
and the resulting .lz4 file is naturally corrupted.
To reproduce:

    mkdir ramdisk
    sudo mount -t tmpfs -o size=1m tmpfs ramdisk/
    ./lz4_compress /some/big/file ramdisk/out.lz4

Discussion:

C++ iostreams were invented before exceptions which is why they are
disabled by default and have to be opted in for each stream.

If exceptions are not enabled, the badbit should be checked before/after
every stream operation. If that is not done, errors go silently
and the user marvels over odd syscall behaviour and corrupt data.

This library does not check the badbit.

This commit fixes the situation it by enabling exceptions using

    stream.exceptions(std::ostream::failbit);

thus enforcing what the library's current assumption that streams
are not in an invalid state.

Further work:

An alternative approach might be to change the library to get rid
of that assumption and switch to the badbit-check everywhere,
but this would be more difficult. It might present good follow-up work
to this fix.

As a reminder:
An application might want to handle ENOSP by looping until more
space is available on the file system, and re-use the open LZ4
stream as soon as that is the case.

This commit still does not make that use case possible, because
if the user catches the exception and calls the library again,
the `input_buffer`/`output_buffer` would be advanced (e.g. by more
LZ4 operations), so the stream would not be valid.

A fix for that would be to observe any stream errors in the library,
and ensure that when the user calls it again, the existing buffers
are not invalidated.